Tool for searching git
As part of the open source discovery at GDS I made a little tool to help
find popular projects on github.

There's a few methods that hep you find how many stars, likes or forks
the repos in an org have.

I need to update the repo to make it clearer how it works.